requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
coverage==6.2
Django==3.2.10
djangorestframework==3.13.1
importlib-metadata==4.9.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
pytest-dependency==0.5.1
pytest-django==4.5.2
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.0
typing_extensions==4.0.1
zipp==3.6.0
